Make it possible to connect row refreshes between instances
By removing observation support in the last commit, one thing we lose is
the ability of multiple grids using the same store to automatically recieve
updates whenever the store is changed by any means (at least, in the
cases where OnDemandList and Observable actually got us that far reliably).
This commit reorganizes some code to make it feasible to connect
the refresh of rows in one grid to another, and perhaps other cases as well.